Add thai data
